feat: export utils with createStorage
