+++ README.md
@@ -243,9 +243,10 @@ This compiled JS files and declarations will be placed in the `dist` folder of t
 
 You can import the followings directly from `@fairdatasociety/beeson`:
 
-* Type          # enum for [types](#Types) used in BeeSon
-* BeeSon        # BeeSon class that you can initialize with either JSON object or TypeSpecification
-* TypeSpecification    # TypeSpecification class that defines JSON object structures/types and its TypeScpecification
+* Type                  # enum for [types](#Types) used in BeeSon
+* BeeSon                # BeeSon class that you can initialize with either JSON object or TypeSpecification
+* TypeSpecification     # TypeSpecification class that defines JSON object structures/types and its TypeScpecification
+* Utils                 # Utility functions such as `createStorage` that can be used for SuperBeeSon handling
 
 Work with non-container types:
 ```js
