+++ src/index.ts
@@ -1,3 +1,7 @@
+import { createStorage } from './utils'
 export * from './beeson'
 export { TypeSpecification } from './type-specification'
 export { Type } from './types'
+export const Utils = {
+  createStorage,
+}
